refactor: 구간 삭제 API 수정
