+++ src/main/java/subway/ui/SectionController.java
@@ -1,16 +1,22 @@
 package subway.ui;
 
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import subway.ui.dto.response.LineResponse;
-import subway.ui.dto.request.SectionRequest;
-import subway.application.SectionService;
-import subway.ui.dto.response.SectionResponse;
-import subway.ui.dto.response.StationResponse;
-
 import java.net.URI;
 import java.util.List;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import subway.application.SectionService;
+import subway.ui.dto.request.SectionDeleteRequest;
+import subway.ui.dto.request.SectionRequest;
+import subway.ui.dto.response.SectionResponse;
+import subway.ui.dto.response.StationResponse;
 
 @RestController
 @RequestMapping("/sections")
@@ -34,8 +40,8 @@ public class SectionController {
         return ResponseEntity.ok(stations);
     }
 
-    @DeleteMapping("/{lineId}/stations")
-    public void deleteSection(@PathVariable(name = "lineId") Long lineId, @RequestBody LineResponse.SectionDeleteRequest sectionDeleteRequest) {
+    @DeleteMapping("/{lineId}")
+    public void deleteSection(@PathVariable(name = "lineId") Long lineId, @RequestBody SectionDeleteRequest sectionDeleteRequest) {
         sectionService.deleteSection(lineId, sectionDeleteRequest);
         ResponseEntity.noContent().build();
     }
